chore(arch): Move lexer and friends into lexer/
