+++ src/index.ts
@@ -1,10 +1,8 @@
 import * as fs from "fs";
 import * as readline from "readline";
 
-import { Token } from "./Token";
-import * as Lexer from "./lexer";
+import { Token, Lexer } from "./lexer";
 import * as Parser from "./parser";
-import { AstPrinter } from "./interpreter/AstPrinter";
 import { Interpreter, OutputStreams } from "./interpreter";
 import { stringify } from "./Stringify";
 import * as BrsError from "./Error";
