+++ src/parser/Statement.ts
@@ -1,6 +1,6 @@
 import * as Expr from "./Expression";
-import { Token } from "../Token";
-import { BrsType, BrsInvalid, Argument, BrsValue, ValueKind } from "../brsTypes";
+import { Token } from "../lexer";
+import { BrsType } from "../brsTypes";
 
 /** A set of reasons why a `Block` stopped executing. */
 export * from "./BlockEndReason";
